Refactor to reduce dependency on external API
By appending new config setting keys to the `@settings` class variable
we were reaching into Nesta's internals in a manner that Nesta might not
have been expecting.

Also, in version 0.14.0, the `@settings` class var has gone away.

The `Nesta::Config` class provides a `fetch()` method that we can use to
read an arbitrary setting, so we can use that instead.

There's also no need to write these values to Sinatra's global settings
object, as they're defined in a scope that makes them available to the
`before` and `after` blocks. So it's a bit less brittle (to unforeseen
changes in Sinatra) to use local variables.